Ca marche

+++ src/main/java/com/example/voiture/Front/Postgres/repository/VoitureRechercheRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.voiture.Front.Postgres.modele.Voiture;
 
-public interface VoitureRechercheRepository extends JpaRepository<Voiture, Long> {
+public interface VoitureRechercheRepository extends JpaRepository<Voiture, Integer> {
     // Vous pouvez ajouter des méthodes spécifiques de requête ici si nécessaire
     Voiture findById(int id);
