+++ src/main/java/com/example/voiture/Mobile/Postgres/repository/HistoEtatRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.voiture.Mobile.Postgres.modele.Histoetat;
 
-public interface HistoEtatRepository extends JpaRepository<Histoetat, Long> {
+public interface HistoEtatRepository extends JpaRepository<Histoetat, Integer> {
     // Vous pouvez ajouter des méthodes spécifiques de requête ici si nécessaire
     List<Histoetat> findByIdvoiture(int id);
 }
